bug: don't have to reload inorder to get to home page
