+++ client/store/index.js
@@ -1,10 +1,11 @@
-import {createStore, applyMiddleware} from 'redux'
+import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import reducer from './reducer'
+import manageUsers from './reducer'
+import manageProducts from './products'
 
-// const reducer = combineReducers({user})
+const reducer = combineReducers({user: manageUsers, products: manageProducts})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
