+++ client/store/reducer.js
@@ -1,17 +1,13 @@
 import axios from 'axios'
 import history from '../history'
-import {combineReducers} from 'redux'
-import Axios from 'axios'
+// import {combineReducers} from 'redux'
 
 /**
  * ACTION TYPES
  */
 const GET_USER = 'GET_USER'
 const REMOVE_USER = 'REMOVE_USER'
-const ALL_PRODUCTS = 'ALL_PRODUCTS'
-const SINGLE_PRODUCT = 'SINGLE_PRODUCT'
 const CREATE_ORDER = 'CREATE_ORDER'
-
 
 /**
  * INITIAL STATE
@@ -23,9 +19,7 @@ const defaultUser = {}
  */
 const getUser = user => ({type: GET_USER, user})
 const removeUser = () => ({type: REMOVE_USER})
-const viewProducts = products => ({type: ALL_PRODUCTS, products})
 const createOrder = order => ({type: CREATE_ORDER, order})
-const singleProduct = product => ({type: SINGLE_PRODUCT, product})
 
 /**
  * THUNK CREATORS
@@ -65,25 +59,10 @@ export const logout = () => async dispatch => {
   }
 }
 
-export const getAllProducts = () => {
-  return async dispatch => {
-    const result = await axios.get('/api/products')
-    dispatch(viewProducts(result.data))
-  }
-}
-
-
 export const purchase = order => {
   return async dispatch => {
     const data = await Axios.post('/api/cart', order)
     dispatch(createOrder(data))
-  }
-}
-
-export const getSingleProduct = product => {
-  return async dispatch => {
-    const result = await axios.get(`/api/products/${product}`)
-    dispatch(singleProduct(result.data))
   }
 }
 
@@ -101,17 +80,6 @@ const manageUsers = (state = defaultUser, action) => {
   }
 }
 
-const manageProducts = (state = [], action) => {
-  switch (action.type) {
-    case ALL_PRODUCTS:
-      return action.products
-    case SINGLE_PRODUCT:
-      return action
-    default:
-      return state
-  }
-}
-
 const manageOrder = (state = [], action) => {
   switch (action.type) {
     case CREATE_ORDER:
@@ -121,9 +89,9 @@ const manageOrder = (state = [], action) => {
   }
 }
 
-const reducer = combineReducers({
-  user: manageUsers,
-  products: manageProducts,
-  order: manageOrder
-})
-export default reducer
+// const reducer = combineReducers({
+//   user: manageUsers,
+//   products: manageProducts,
+//   order: manageOrder
+// })
+export default manageUsers
